feat: 增加错误校验

+++ dynamic-thread-pool-spring-boot-stater/src/main/java/cn/comradexy/middleware/sdk/registry/redis/RedisRegistry.java
@@ -6,6 +6,8 @@ import cn.comradexy.middleware.sdk.registry.IRegistry;
 import org.redisson.api.RBucket;
 import org.redisson.api.RList;
 import org.redisson.api.RedissonClient;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.time.Duration;
 import java.util.List;
@@ -18,6 +20,8 @@ import java.util.List;
  * @Description: Redis 注册中心
  */
 public class RedisRegistry implements IRegistry {
+    private final Logger logger = LoggerFactory.getLogger(RedisRegistry.class);
+
     private final RedissonClient redissonClient;
 
     public RedisRegistry(RedissonClient redissonClient) {
@@ -26,20 +30,27 @@ public class RedisRegistry implements IRegistry {
 
     @Override
     public void reportThreadPool(List<ThreadPoolConfigEntity> threadPoolEntities) {
-        // TODO: 上报所有线程池信息
         RList<ThreadPoolConfigEntity> list =
                 redissonClient.getList(RegistryEnumVO.THREAD_POOL_CONFIG_LIST_KEY.getKey());
+        if (null == list) {
+            logger.error("Redis 注册中心，获取缓存对象失败。{}", RegistryEnumVO.THREAD_POOL_CONFIG_LIST_KEY.getKey());
+            return;
+        }
         list.delete();
         list.addAll(threadPoolEntities);
     }
 
     @Override
     public void reportThreadPoolConfigParameter(ThreadPoolConfigEntity threadPoolConfigEntity) {
-        // TODO: 上报单个线程池的配置参数
         String cacheKey = RegistryEnumVO.THREAD_POOL_CONFIG_PARAMETER_LIST_KEY.getKey() +
                 "_" + threadPoolConfigEntity.getAppName() +
                 "_" + threadPoolConfigEntity.getThreadPoolName();
         RBucket<ThreadPoolConfigEntity> bucket = redissonClient.getBucket(cacheKey);
+        if (null == bucket) {
+            logger.error("Redis 注册中心，获取缓存对象失败。{}", cacheKey);
+            return;
+        }
+        // 缓存30天
         bucket.set(threadPoolConfigEntity, Duration.ofDays(30));
     }
